# Speed up per token and per tensor quant by 15%