<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
#2965 

Improvements
1. Store using `uint4` cast to store 1 time for 16 fp8. Originally it stores 16 times.
2. Increase vec size so it's 16 fp8 -> align with 128 bit
3. Fix per tensor AMD error

Per Token
```
original
✅ All implementations match
per-token-dynamic-quant-fp8-performance:
    batch_size  seq_len         VLLM   SGL Kernel
0         16.0     64.0    26.432000    25.024001
1         16.0    128.0    44.640001    37.535999
2         16.0    256.0    83.632007    68.415999
3         16.0    512.0   154.592007   119.360000
4         16.0   1024.0   296.128005   219.920009
5         16.0   2048.0   579.584002   422.080010
6         16.0   4096.0  1148.480058   830.143988
7         32.0     64.0    44.544000    38.368002
8         32.0    128.0    84.639996    68.480000
9         32.0    256.0   155.456007   119.231999
10        32.0    512.0   296.256006   220.128000
11        32.0   1024.0   578.880012   422.176003
12        32.0   2048.0  1148.192048   830.175996
13        32.0   4096.0  2280.319929  1637.552023
14        64.0     64.0    84.639996    69.568001
15        64.0    128.0   155.520007   119.935997
16        64.0    256.0   296.128005   220.063999
17        64.0    512.0   578.815997   421.759993
18        64.0   1024.0  1148.095965   829.792023
19        64.0   2048.0  2280.352116  1637.151957
20        64.0   4096.0  4541.728020  3251.199961
21       128.0     64.0   155.472010   119.424000
22       128.0    128.0   296.319991   220.543995
23       128.0    256.0   578.688025   422.592014
24       128.0    512.0  1148.512006   830.048025
25       128.0   1024.0  2279.664040  1637.311935
26       128.0   2048.0  4544.159889  3253.103971
27       128.0   4096.0  9106.351852  6487.264156

write once
✅ All implementations match
per-token-dynamic-quant-fp8-performance:
    batch_size  seq_len         VLLM   SGL Kernel
0         16.0     64.0    26.176000    22.304000
1         16.0    128.0    44.351999    35.232000
2         16.0    256.0    83.424002    64.015999
3         16.0    512.0   154.336005   109.824002
4         16.0   1024.0   295.711994   198.975995
5         16.0   2048.0   581.503987   378.208011
6         16.0   4096.0  1154.144049   739.968002
7         32.0     64.0    44.287998    34.175999
8         32.0    128.0    84.639996    64.191997
9         32.0    256.0   155.744001   109.696001
10        32.0    512.0   297.087997   199.231997
11        32.0   1024.0   580.896020   378.304005
12        32.0   2048.0  1153.759956   740.064025
13        32.0   4096.0  2289.247990  1456.383944
14        64.0     64.0    84.671997    63.072003
15        64.0    128.0   155.775994   109.856002
16        64.0    256.0   297.087997   199.328005
17        64.0    512.0   581.055999   377.119988
18        64.0   1024.0  1152.608037   739.887953
19        64.0   2048.0  2290.623903  1456.336021
20        64.0   4096.0  4562.208176  2890.496016
21       128.0     64.0   155.775994   108.800001
22       128.0    128.0   297.087997   198.336005
23       128.0    256.0   580.896020   377.023995
24       128.0    512.0  1152.608037   740.095973
25       128.0   1024.0  2289.440155  1456.383944
26       128.0   2048.0  4563.839912  2890.559912
27       128.0   4096.0  9104.880333  5753.727913

change to load 16 items
✅ All implementations match
per-token-dynamic-quant-fp8-performance:
    batch_size  seq_len         VLLM   SGL Kernel
0         16.0     64.0    26.400000    23.232000
1         16.0    128.0    44.895999    32.960001
2         16.0    256.0    83.903998    58.944002
3         16.0    512.0   155.200005   100.736000
4         16.0   1024.0   297.248006   186.240003
5         16.0   2048.0   581.856012   355.455995
6         16.0   4096.0  1154.416084   697.712004
7         32.0     64.0    44.576000    34.175999
8         32.0    128.0    84.831998    58.912002
9         32.0    256.0   156.000003   101.728000
10        32.0    512.0   297.408015   186.271995
11        32.0   1024.0   581.215978   355.520010
12        32.0   2048.0  1154.047966   697.791994
13        32.0   4096.0  2290.719986  1374.448061
14        64.0     64.0    84.927998    57.567999
15        64.0    128.0   156.000003   100.671999
16        64.0    256.0   297.280014   185.376003
17        64.0    512.0   581.215978   354.528010
18        64.0   1024.0  1153.087974   697.664022
19        64.0   2048.0  2289.920092  1374.688029
20        64.0   4096.0  4563.712120  2730.432034
21       128.0     64.0   155.936003   101.792000
22       128.0    128.0   297.376007   186.335996
23       128.0    256.0   581.344008   355.360001
24       128.0    512.0  1154.047966   697.727978
25       128.0   1024.0  2291.039944  1375.056028
26       128.0   2048.0  4563.456059  2730.175972
27       128.0   4096.0  9105.728149  5436.592102
```
Per Tensor
```
original
✅ All implementations match
per-tensor-quant-fp8-performance:
    batch_size  seq_len         VLLM   SGL Kernel
0         16.0     64.0    37.023999    26.752001
1         16.0    128.0    60.736001    39.840002
2         16.0    256.0   119.231999    73.119998
3         16.0    512.0   232.352003   146.559998
4         16.0   1024.0   443.168014   273.808002
5         16.0   2048.0   861.999989   523.712039
6         32.0     64.0    60.608000    39.840002
7         32.0    128.0   119.935997    72.959997
8         32.0    256.0   232.575998   147.775993
9         32.0    512.0   443.071991   274.304003
10        32.0   1024.0   861.631989   524.544001
11        32.0   2048.0  1703.328013  1022.431970
12        64.0     64.0   119.999997    72.991997
13        64.0    128.0   232.544005   147.471994
14        64.0    256.0   443.040013   273.920000
15        64.0    512.0   861.616015   524.223983
16        64.0   1024.0  1707.008004  1023.231983
17        64.0   2048.0  3376.847982  2017.055988
18       128.0     64.0   232.495993   147.872001
19       128.0    128.0   443.136007   274.079978
20       128.0    256.0   861.328006   523.952007
21       128.0    512.0  1706.303954  1024.512053
22       128.0   1024.0  3376.512051  2011.967897
23       128.0   2048.0  6726.352215  3974.447966


after tune
✅ All implementations match
per-tensor-quant-fp8-performance:
    batch_size  seq_len         VLLM   SGL Kernel
0         16.0     64.0    37.055999    26.112000
1         16.0    128.0    60.864002    37.983999
2         16.0    256.0   119.840004    69.408000
3         16.0    512.0   232.575998   128.128007
4         16.0   1024.0   443.551987   234.255999
5         16.0   2048.0   861.792028   449.151993
6         32.0     64.0    60.736001    38.047999
7         32.0    128.0   119.903997    69.263995
8         32.0    256.0   232.767999   128.031999
9         32.0    512.0   443.264008   234.320000
10        32.0   1024.0   861.728013   449.151993
11        32.0   2048.0  1703.855991   880.367994
12        64.0     64.0   120.127998    69.087997
13        64.0    128.0   232.575998   127.952009
14        64.0    256.0   443.232000   234.336004
15        64.0    512.0   861.487985   449.119985
16        64.0   1024.0  1706.496000   880.352020
17        64.0   2048.0  3377.871990  1735.872030
18       128.0     64.0   232.703999   127.936006
19       128.0    128.0   443.423986   234.336004
20       128.0    256.0   861.904025   449.119985
21       128.0    512.0  1706.815958   880.735993
22       128.0   1024.0  3376.751900  1735.072017
23       128.0   2048.0  6725.279808  3447.567940
```

Profiling showed inefficient storing

<img width="1338" alt="Screenshot 2025-03-19 at 9 32 40 AM" src="https://github.com/user-attachments/assets/92e6d2bb-4a20-464b-ae71-25d98705f41f" />

After storage and vec_size changes:
![Screenshot 2025-03-20 at 2 38 09 PM](https://github.com/user-attachments/assets/a4299c78-91b2-4ab3-9142-dd351a0509b3)


## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.